# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Nexvault — Digital Banking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /onboarding/start:
    post:
      operationId: startOnboarding
      summary: Start onboarding flow
      tags: [onboarding]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartOnboardingRequest"
      responses:
        "200":
          description: Onboarding started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingFlowResponse"
  /onboarding/step:
    post:
      operationId: submitOnboardingStep
      summary: Submit onboarding step
      tags: [onboarding]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingStepRequest"
      responses:
        "200":
          description: Step submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingStepResponse"
...
  /payments/beneficiaries:
    get:
      operationId: getBeneficiaries
      summary: List beneficiaries
      tags: [payments]
      responses:
        "200":
          description: Beneficiaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Beneficiary"
    post:
      operationId: createBeneficiary
      summary: Add beneficiary
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBeneficiaryRequest"
      responses:
        "201":
          description: Beneficiary added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Beneficiary"
  /payments/fx/quote:
    get:
      operationId: getFxQuote
      summary: Get FX quote
      tags: [payments]
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
        - name: amount
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: FX quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FxQuote"
  /payments/transfer/initiate:
    post:
      operationId: initiateTransfer
      summary: Initiate transfer
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateTransferRequest"
      responses:
        "200":
          description: Transfer initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferInitiateResponse"
  /payments/transfer/confirm:
    post:
      operationId: confirmTransfer
      summary: Confirm transfer with OTP
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmTransferRequest"
      responses:
        "200":
          description: Transfer confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/transfers:
    get:
      operationId: getTransfers
      summary: Transfer history
      tags: [payments]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Transfers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentList"

  # Notifications
...
components:
  schemas:
...
    StartOnboardingRequest:
      type: object
      required: [flowType, countryCode]
      properties:
        flowType:
          type: string
          enum: [individual, business]
        countryCode:
          type: string
    OnboardingFlowResponse:
      type: object
      required: [applicationId, currentStep, flowConfig]
      properties:
        applicationId:
          type: string
        currentStep:
          type: string
        flowConfig:
          type: object
          additionalProperties: true
    OnboardingStepRequest:
      type: object
      required: [applicationId, stepId, data]
      properties:
        applicationId:
          type: string
        stepId:
          type: string
        data:
          type: object
          additionalProperties: true
    OnboardingStepResponse:
      type: object
      required: [stepId, nextStep, status]
      properties:
        stepId:
          type: string
        nextStep:
          type: string
        status:
          type: string
          enum: [in_progress, kyc_pending, approved, rejected]
...
    Beneficiary:
      type: object
      required: [id, name, currency, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        accountNumber:
          type: string
        sortCode:
          type: string
        iban:
          type: string
        swiftBic:
          type: string
        bankName:
          type: string
        country:
          type: string
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateBeneficiaryRequest:
      type: object
      required: [name, currency]
      properties:
        name:
          type: string
        accountNumber:
          type: string
        sortCode:
          type: string
        iban:
          type: string
        swiftBic:
          type: string
        bankName:
          type: string
        country:
          type: string
        currency:
          type: string
    FxQuote:
      type: object
      required: [from, to, rate, amount, convertedAmount, fee, expiresAt]
      properties:
        from:
          type: string
        to:
          type: string
        rate:
          type: number
        amount:
          type: number
        convertedAmount:
          type: number
        fee:
          type: number
        quoteId:
          type: string
        expiresAt:
          type: string
          format: date-time
    InitiateTransferRequest:
      type: object
      required: [fromAccountId, beneficiaryId, amount, currency]
      properties:
        fromAccountId:
          type: string
        beneficiaryId:
          type: string
        amount:
          type: number
        currency:
          type: string
        reference:
          type: string
        quoteId:
          type: string
    TransferInitiateResponse:
      type: object
      required: [transferId, requiresOtp]
      properties:
        transferId:
          type: string
        requiresOtp:
          type: boolean
        fee:
          type: number
        estimatedArrival:
          type: string
    ConfirmTransferRequest:
      type: object
      required: [transferId, otpCode]
      properties:
        transferId:
          type: string
        otpCode:
          type: string
    Payment:
      type: object
      required: [id, amount, currency, status, createdAt]
      properties:
        id:
          type: string
        fromAccountId:
          type: string
        beneficiaryId:
          type: string
        amount:
          type: number
        sourceCurrency:
          type: string
        destinationCurrency:
          type: string
        exchangeRate:
          type: number
        destinationAmount:
          type: number
        fee:
          type: number
        currency:
          type: string
        reference:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        providerReference:
          type: string
        initiatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    PaymentList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
...