# lib/api-zod/src/generated/types/fxQuote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexvault — Digital Banking Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface FxQuote {
  from: string;
  to: string;
  rate: number;
  amount: number;
  convertedAmount: number;
  fee: number;
  quoteId?: string;
  expiresAt: Date;
}