# lib/api-zod/src/generated/types/onboardingStepRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexvault — Digital Banking Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { OnboardingStepRequestData } from "./onboardingStepRequestData";

export interface OnboardingStepRequest {
  applicationId: string;
  stepId: string;
  data: OnboardingStepRequestData;
}