# lib/api-zod/src/generated/types/onboardingStepResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexvault — Digital Banking Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { OnboardingStepResponseStatus } from "./onboardingStepResponseStatus";

export interface OnboardingStepResponse {
  stepId: string;
  nextStep: string;
  status: OnboardingStepResponseStatus;
}