# lib/api-zod/src/generated/types/paymentList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexvault — Digital Banking Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { Payment } from "./payment";

export interface PaymentList {
  items: Payment[];
  total: number;
}